Progress bar, change grade key front end
Progress bar front end is done, will eventually be connected to AI model.

Change grade key pops up file explorer and allows user to select a new file that must be of type png,jpeg,heic, or .jpg. It will change the grade key viewer on the right side of the app if the file chosen is of the supported types

Will pop up with an error message if it is not of the supported types until a supported type is chosen