Prototype 2 code and sample .png
This code is the second prototype that works directly with a Google Sheet that is shared with the team. 
Acceptance criteria: the new prototype should be able to read a specific column of cells that contain an image of homework, and it should then be able to print out (to another specified cell) what was written on the assignment. Currently no grading function is implemented, just picture to text generation.