Fixed bugs and refactored detect_blur
detect_blur now works on a per image basis and returns true or false so that it can be easily integrated with existing framework and also so it can easily be stopped. Fixed issues with data loader that would not allow for training if there were missing classes on an image.